% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_ctlars_utils.R
\docType{package}
\name{ctlars-package}
\alias{ctlars-package}
\title{ctlars: ctlars - Complex Terminating Lars Algorithm}
\description{
The package implements a early terminating version of the complex Lars algorithm.
}
\details{
ctlars is a package showing how to produce packages with
Rcpp Modules and C++ classes and make them available under S4.
}
\author{
\strong{Maintainer}: Fabian Scheidt \email{fabian.scheidt@tu-darmstadt.de}

Authors:
\itemize{
  \item Jasin Machkour \email{jasin.machkour@tu-darmstadt.de}
  \item Michael Muma \email{michael.muma@tu-darmstadt.de}
}

}
