% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctlars.R
\docType{class}
\name{ctlars}
\alias{ctlars}
\title{ctlars R6 class.}
\description{
The implementation of the ctlars algorithm as R6 class.
}
\examples{
set.seed(42)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{x}}{The predictor matrix.}

\item{\code{y}}{The response vector.}

\item{\code{has_intercept}}{Logical. Whether the model includes an intercept.}

\item{\code{standardize}}{Logical. Whether to standardize the predictor matrix.}

\item{\code{verbose}}{Logical. Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ctlars-new}{\code{ctlars$new()}}
\item \href{#method-ctlars-execute_clars_step}{\code{ctlars$execute_clars_step()}}
\item \href{#method-ctlars-get_active_set}{\code{ctlars$get_active_set()}}
\item \href{#method-ctlars-get_beta_hat}{\code{ctlars$get_beta_hat()}}
\item \href{#method-ctlars-get_beta_history}{\code{ctlars$get_beta_history()}}
\item \href{#method-ctlars-get_y_hat}{\code{ctlars$get_y_hat()}}
\item \href{#method-ctlars-get_ssr}{\code{ctlars$get_ssr()}}
\item \href{#method-ctlars-get_r2}{\code{ctlars$get_r2()}}
\item \href{#method-ctlars-get_active_dummies}{\code{ctlars$get_active_dummies()}}
\item \href{#method-ctlars-get_num_dummies}{\code{ctlars$get_num_dummies()}}
\item \href{#method-ctlars-get_n}{\code{ctlars$get_n()}}
\item \href{#method-ctlars-get_p}{\code{ctlars$get_p()}}
\item \href{#method-ctlars-clone}{\code{ctlars$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-new"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-new}{}}}
\subsection{Method \code{new()}}{
Create a new ctlars object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$new(
  x,
  y,
  has_intercept = TRUE,
  standardize = TRUE,
  num_dummies = ncol(x),
  verbose = TRUE,
  tol = .Machine$double.eps
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The predictor matrix.}

\item{\code{y}}{The response vector.}

\item{\code{has_intercept}}{Logical. Whether the model includes an intercept. Default is TRUE.}

\item{\code{standardize}}{Logical. Whether to standardize the predictor matrix. Default is TRUE.}

\item{\code{num_dummies}}{Integer. Number of dummy variables. Default is the number of columns in \code{x}.}

\item{\code{verbose}}{Logical. Whether to print progress messages. Default is TRUE.}

\item{\code{tol}}{Numeric. Tolerance for numerical calculations. Default is \code{.Machine$double.eps}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new ctlars object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-execute_clars_step"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-execute_clars_step}{}}}
\subsection{Method \code{execute_clars_step()}}{
Execute a CLARS step.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$execute_clars_step(t_stop, early_stop = TRUE, use_chol = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_stop}}{Numeric. Stopping criterion for the algorithm.}

\item{\code{early_stop}}{Logical. Whether to stop early based on certain conditions. Default is TRUE.}

\item{\code{use_chol}}{Logical. Whether to use Cholesky decomposition. Default is TRUE.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_active_set"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_active_set}{}}}
\subsection{Method \code{get_active_set()}}{
Get the active set of predictors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_active_set()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of indices representing the active set of predictors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_beta_hat"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_beta_hat}{}}}
\subsection{Method \code{get_beta_hat()}}{
Get the estimated coefficients.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_beta_hat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of estimated coefficients.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_beta_history"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_beta_history}{}}}
\subsection{Method \code{get_beta_history()}}{
Get the history of estimated coefficients.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_beta_history()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A matrix of estimated coefficients over iterations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_y_hat"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_y_hat}{}}}
\subsection{Method \code{get_y_hat()}}{
Get the predicted response values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_y_hat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of predicted response values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_ssr"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_ssr}{}}}
\subsection{Method \code{get_ssr()}}{
Get the sum of squared residuals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_ssr()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric value representing the sum of squared residuals.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_r2"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_r2}{}}}
\subsection{Method \code{get_r2()}}{
Get the R-squared value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_r2()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric value representing the R-squared value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_active_dummies"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_active_dummies}{}}}
\subsection{Method \code{get_active_dummies()}}{
Get the active dummy variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_active_dummies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A vector of indices representing the active dummy variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_num_dummies"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_num_dummies}{}}}
\subsection{Method \code{get_num_dummies()}}{
Get the number of dummy variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_num_dummies()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer representing the number of dummy variables.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_n"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_n}{}}}
\subsection{Method \code{get_n()}}{
Get the number of observations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_n()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer representing the number of observations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-get_p"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-get_p}{}}}
\subsection{Method \code{get_p()}}{
Get the number of predictors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$get_p()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer representing the number of predictors.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ctlars-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ctlars-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ctlars$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
