% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDP.R
\name{FDP}
\alias{FDP}
\title{False Detection Proportion}
\usage{
FDP(selected_vars, true_actives)
}
\arguments{
\item{selected_vars}{a vector with estimated support indices.}

\item{true_actives}{a vector with true support indices.}
}
\value{
Returns the false detection proportion as a number.
}
\description{
The function returns the empirical false detection proportion based on a
vector of selected variable indices and a vector of reference indices.
}
\examples{
set.seed(42)
}
