% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ctlars.R
\name{plot.ctlars}
\alias{plot.ctlars}
\title{Plots the CT-LARS solution path.}
\usage{
\method{plot}{ctlars}(
  x,
  xlab = "# Included dummies",
  ylab = "|Coefficients|",
  include_dummies = TRUE,
  actions = TRUE,
  col_selected = "black",
  col_dummies = "red",
  lty_selected = "solid",
  lty_dummies = "dashed",
  legend_pos = "topleft",
  ...
)
}
\arguments{
\item{x}{Object of the class ctlars See \link{ctlars} for details.}

\item{xlab}{Label of the x-axis.}

\item{ylab}{Label of the y-axis.}

\item{include_dummies}{Logical. If TRUE solution paths of dummies are added to the plot.}

\item{actions}{Logical. If TRUE axis above plot with indices of added variables
(Dummies represented by 'D') along the solution path is added.}

\item{col_selected}{Color of lines corresponding to selected variables.}

\item{col_dummies}{Color of lines corresponding to included dummies.}

\item{lty_selected}{Line type of lines corresponding to selected variables.
See \link{par} for more details.}

\item{lty_dummies}{Line type of lines corresponding to included dummies.
See \link{par} for more details.}

\item{legend_pos}{Legend position. See \link{xy.coords} for more details.}

\item{...}{Ignored. Only added to keep structure of generic \link{plot} function.}
}
\value{
Plots the ctlars solution path stored in a ctlars object.
}
\description{
Plots the CT-LARS solution path.
(see \link{ctlars} for details.)
}
\examples{
set.seed(42)
}
\seealso{
\link{ctlars}, \link{plot}, \link{par}, and \link{xy.coords}.
}
