% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_model_TPR_FDR.R
\name{simulate_model_TPR_FDR}
\alias{simulate_model_TPR_FDR}
\title{Simulate a ctlars model with True Positive Rate and False Detection Rate}
\usage{
simulate_model_TPR_FDR(
  T_vec = c(1, 2, 5, 10, 20, 50, 100),
  MC = 2,
  n = 10,
  p = 30,
  num_actives = 3,
  num_dummies = 30,
  noise_power = 1,
  set_snr = FALSE,
  snr_is_linear = TRUE,
  snr_val_linear = 2,
  snr_val_db = 10,
  verbose = TRUE
)
}
\arguments{
\item{T_vec}{Vector of terminating steps to conduct the simulation.
Default is c(1, 2, 5, 10, 20, 50, 100).}

\item{MC}{Number of Monte Carlo experiments per termination step.
Default is 50.}

\item{n}{Number of observations/rows for the Gaussian regression.
Default is 100.}

\item{p}{Number of predictors/columns for the Gaussian regression.
Default is 300.}

\item{num_actives}{Number of active variables used in Gaussian regression.
Default is 10.}

\item{num_dummies}{Number of dummy variables added to the predictor matrix.
Default is \code{p}.}

\item{noise_power}{Numeric value indicating the noise power of white
centered circulary symmetric Gaussian noise. Default is 1.
Can be computed by setting the parameter \code{set_snr} to \code{TRUE}.}

\item{set_snr}{Logical indicating if SNR (Signa-to-Noise Ratio) is controlled.
Default is FALSE.}

\item{snr_is_linear}{Logical indicating if SNR (Signal-to-Noise Ratio) value
is given in linear scale (\code{TRUE}) or dB scale (\code{FALSE}). Default is \code{TRUE}.}

\item{snr_val_linear}{SNR value in linear scale. Default is 2.}

\item{snr_val_db}{SNR value in dB scale. Default is 10.}

\item{verbose}{Logical indicating whether a simulation progress is displayed.
Default is \code{TRUE}.}
}
\value{
The function returns a list with the following components:
\itemize{
\item{T_vec} Vector with realized terminating steps.
\item{FDP} Matrix with realized False Discovery Proportions (FDP).
\item{TPP} Matrix with realized True Positive Proportions (TPP).
\item{FDR} Vector with realized False Discovery Rates (FDR).
\item{TPR} Vector with realized True Positive Rates (TPR).
}
}
\description{
From a statistical point of view, it is desirable to use a variable selection
method that allows for controlling the expected value of the FDP at a user
defined target level alpha in \code{[}0, 1\verb{]}, while maximizing the number of selected
variables.
These type of methods exist and are called false discovery rate (FDR)-
controlling methods.
For example, the CT-Rex Selector is a fast and FDR-controlling variable/feature
selection framework for large-scale high-dimensional settings that relies on
the ctars method.

Consider the following definitions of FDR and TPR:
Let AA be part of \verb{\{1, ..., p\}} be the set of selected variables, and let
\code{A} in \verb{\{1, ..., p\}} be the set of true active variables, \verb{|AA|} be the cardinality
of \code{AA}, and define \verb{1 v a := max\{1, a\}}, a in R.
Then, the false discovery rate (FDR) and the true positive rate (TPR) are
defined by:

\verb{FDR := E[FDP] := E[ |[AA\\A]|/(1 v |AA|) ]}
\verb{TPR := E[TPP] := E[ |[intersection(A, AA)]|/(1 v |A|) ]}

This function simulates a Gaussian regression scenario with complex numbers,
applies the ctlars method for variable selection, and computes FDR and TPR
metrics across multiple termination steps. It generates noisy observations
and includes a specified number of active and dummy variables.
}
\examples{
set.seed(42)
result <- simulate_model_TPR_FDR()
print(result)
}
\seealso{
\code{\link{ctlars}} for details on the ctlars algorithm.
\code{\link{FDP}} for details on calculating False Discovery Proportion.
\code{\link{TPP}} for details on calculating True Positive Proportion.
}
