% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPP.R
\name{TPP}
\alias{TPP}
\title{True Detection Proportion}
\usage{
TPP(selected_vars, true_actives)
}
\arguments{
\item{selected_vars}{a vector with estimated support indices.}

\item{true_actives}{a vector with true support indices.}
}
\value{
Returns the true positive proporiton as a number.
}
\description{
The function returns the empirical true detection proportion based on a
vector of selected variable indices and a vector of reference indices.
}
\examples{
set.seed(42)
}
