% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCCCGauss_data_small.R
\docType{data}
\name{WCCCGauss_data_small}
\alias{WCCCGauss_data_small}
\title{WCCCGauss Data small}
\format{
A list with the following components:
\describe{
\item{X}{A complex matrix with 10 rows and 15 columns, containing complex numbers.}
\item{beta}{A complex vector of length 15.}
\item{n}{An integer representing the number of observations, which is 10.}
\item{p}{An integer representing the number of variables, which is 15.}
\item{support}{An integer vector indicating the support indices.}
\item{y}{A complex matrix with one column and 10 rows.}
\item{beta_card}{An integer representing the cardinality of beta, which is 5.}
\item{rnd_state}{An integer representing the random state, which is 2.}
}
}
\usage{
data(WCCCGauss_data_small)
}
\description{
A dataset containing complex numbers and related parameters for testing
the CTLARS algorithm.
}
\keyword{datasets}
